+++ tests/integration/api/v2010/account/test_message.py
@@ -153,9 +153,7 @@ class MessageTestCase(IntegrationTestCase):
             200,
             '''
             {
-                "end": 0,
                 "first_page_uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages.json?PageSize=1&Page=0",
-                "last_page_uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages.json?PageSize=1&Page=119771",
                 "messages": [
                     {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
@@ -183,12 +181,9 @@ class MessageTestCase(IntegrationTestCase):
                     }
                 ],
                 "next_page_uri": null,
-                "num_pages": 119772,
                 "page": 0,
                 "page_size": 1,
                 "previous_page_uri": null,
-                "start": 0,
-                "total": 119772,
                 "uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages.json?PageSize=1&Page=0"
             }
             '''