+++ twilio/rest/api/v2010/account/sip/ip_access_control_list/ip_address.py
@@ -118,17 +118,22 @@ class IpAddressList(ListResource):
 
         return IpAddressPage(self._version, response, self._solution)
 
-    def create(self, friendly_name, ip_address):
+    def create(self, friendly_name, ip_address, cidr_prefix_length=values.unset):
         """
         Create a new IpAddressInstance
 
         :param unicode friendly_name: The friendly_name
         :param unicode ip_address: The ip_address
+        :param unicode cidr_prefix_length: The cidr_prefix_length
 
         :returns: Newly created IpAddressInstance
         :rtype: twilio.rest.api.v2010.account.sip.ip_access_control_list.ip_address.IpAddressInstance
         """
-        data = values.of({'FriendlyName': friendly_name, 'IpAddress': ip_address, })
+        data = values.of({
+            'FriendlyName': friendly_name,
+            'IpAddress': ip_address,
+            'CidrPrefixLength': cidr_prefix_length,
+        })
 
         payload = self._version.create(
             'POST',
@@ -279,17 +284,23 @@ class IpAddressContext(InstanceContext):
             sid=self._solution['sid'],
         )
 
-    def update(self, ip_address=values.unset, friendly_name=values.unset):
+    def update(self, ip_address=values.unset, friendly_name=values.unset,
+               cidr_prefix_length=values.unset):
         """
         Update the IpAddressInstance
 
         :param unicode ip_address: The ip_address
         :param unicode friendly_name: The friendly_name
+        :param unicode cidr_prefix_length: The cidr_prefix_length
 
         :returns: Updated IpAddressInstance
         :rtype: twilio.rest.api.v2010.account.sip.ip_access_control_list.ip_address.IpAddressInstance
         """
-        data = values.of({'IpAddress': ip_address, 'FriendlyName': friendly_name, })
+        data = values.of({
+            'IpAddress': ip_address,
+            'FriendlyName': friendly_name,
+            'CidrPrefixLength': cidr_prefix_length,
+        })
 
         payload = self._version.update(
             'POST',
@@ -344,6 +355,7 @@ class IpAddressInstance(InstanceResource):
             'account_sid': payload['account_sid'],
             'friendly_name': payload['friendly_name'],
             'ip_address': payload['ip_address'],
+            'cidr_prefix_length': deserialize.integer(payload['cidr_prefix_length']),
             'ip_access_control_list_sid': payload['ip_access_control_list_sid'],
             'date_created': deserialize.rfc2822_datetime(payload['date_created']),
             'date_updated': deserialize.rfc2822_datetime(payload['date_updated']),
@@ -409,6 +421,14 @@ class IpAddressInstance(InstanceResource):
         return self._properties['ip_address']
 
     @property
+    def cidr_prefix_length(self):
+        """
+        :returns: An integer representing the length of the CIDR prefix to use with this IP address when accepting traffic. By default the entire IP address is used.
+        :rtype: unicode
+        """
+        return self._properties['cidr_prefix_length']
+
+    @property
     def ip_access_control_list_sid(self):
         """
         :returns: The ip_access_control_list_sid
@@ -449,17 +469,23 @@ class IpAddressInstance(InstanceResource):
         """
         return self._proxy.fetch()
 
-    def update(self, ip_address=values.unset, friendly_name=values.unset):
+    def update(self, ip_address=values.unset, friendly_name=values.unset,
+               cidr_prefix_length=values.unset):
         """
         Update the IpAddressInstance
 
         :param unicode ip_address: The ip_address
         :param unicode friendly_name: The friendly_name
+        :param unicode cidr_prefix_length: The cidr_prefix_length
 
         :returns: Updated IpAddressInstance
         :rtype: twilio.rest.api.v2010.account.sip.ip_access_control_list.ip_address.IpAddressInstance
         """
-        return self._proxy.update(ip_address=ip_address, friendly_name=friendly_name, )
+        return self._proxy.update(
+            ip_address=ip_address,
+            friendly_name=friendly_name,
+            cidr_prefix_length=cidr_prefix_length,
+        )
 
     def delete(self):
         """